sgx: add new plugin.
The SGX plugin exposes two device files as separate resources:

  * /dev/sgx/enclave   as sgx.intel.com/enclave
  * /dev/sgx/provision as sgx.intel.com/provision

The number of resources is configurable, but it's intended to be equal
to the pod count by default, so that any pod requiring access would have
it. The access control (who can do SGX remote attestation) is done
outside this plugin.

Signed-off-by: Ismo Puustinen <ismo.puustinen@intel.com>