gpu: log when number of devices found changes
The plugin did not report how many devices it found in the logs.
For a 'normal' case, where devices were found, this was probably not
a big deal, as the devices get advertised as resources on the node.
In the case where no devices are found however, no devices get
registered on the node, and there is no indication in the logs that
nothing was found to register.

Improve this by logging when the number of devices found changes,
which includes logging for '0 devices' in the first instance.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>