ci: fix docs publishing
The initial publish.yml was a merger of two examples with
different ways to handle the publishing and resulted in
some inconsistencies.

v2 of the docs publishing correctly takes the git author
info from the original respository and uses that info to
create the docs-only repo ($HOME/output) that is then pushed
to gh-pages.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>