ci: update Ubuntu versions
In Jenkins, we are using Xenial (16.04) as the worker distro and
that is causing some maintenance issues. For instance, the manually
installed buildah version lacks behind Github's CI that gets
buildah from a repository.

This change moves Jenkins to Ubuntu 18.04 so we get to install
buildah from packages directly too.

While we are at it, move Github VMs to Ubuntu 20.04. This allows
us the change with both LTS versions.

Finally, increase Jenkins timeout from 3 to 4 hours to be able to
build all the demo container images without getting aborted.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>