images: update license obligations
In Debian, uuid and uuid-dev are totally different things.
accel-config build depends on uuid-dev headers but the runtime
library is libuuid1. This is part of debian:sid-slim so we can
just drop installing uuid. Moreover, it's sufficient that we
only ship accel-config source.

Also, correctly ship OPAE SDK COPYING in the final image.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>