ci: use buildah from the distro
After the following announcement on podman.io/blogs:

"Ubuntu 22.04 LTS Beta is available for testing as of March 31st. This
is the first LTS release with Podman, Buildah and Skopeo in the default
repos...

[snip]

There won’t be any further updates to the Kubic repos as far as Podman,
Buildah and Skopeo are concerned, so users are recommended to use the
default repos on 22.04 LTS."

buildah disappeared from the apt archive we were using making CI to
fail. Fortunately, we now get to use buildah from the worker directly.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>