fpga,dsa,iaa: Fix CI issue
Due to a glibc version mismatch in the debian repos and the
unstable-slim and testing-slim docker image 3 CI jobs have been failing
for some time. The issue was fixed by switching to testing-slim base image
and started again when the bug was introduced to the image.

Using testing-slim image as a base should fix the issue.

Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>