e2e: ensure device plugin gets deleted before next one is deployed
Though namespace is deleted after each It(), it does not ensure
that it is deleted. Because of this reason, device plugin did not
get deleted before the next one is deployed. This can cause a
temporary crash of the new plugin and sometimes becomes the cause
of e2e test's failure. This commit fixes it by ensuring previous
device plugin gets deleted after each run.

Signed-off-by: Hyeongju Johannes Lee <hyeongju.lee@intel.com>