pkg/topology: get sysfs parsing from containers/nri-plugins
A proper upstream for the sysfs topology parsing is
containers/nri-plugins se we can drop the duplicated code.

At the same time, we no longer need to carry scripts/ttar
and the sys.ttar unpacking for the tests because the remaining
testdata becomes easier to maintain (the original testdata
contained PCI BDF paths that trigged issues).

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>