e2e: use JustBeforeEach for deploying plugins
There can be some cases that something should be done before
deploying a plugin (e.g. creation of a configMap). To make plugin
gets deployed after some other works are done only in some
'Context', we need to make plugins deployed in 'JustBeforeEach' so
that 'BeforeEach' can run first even if it is inside 'Context'.

Signed-off-by: Hyeongju Johannes Lee <hyeongju.lee@intel.com>