e2e: fix DLB testing
We have suffered from intermitted lockups on GNR targets. This
masked results from e2e-dlb and a broken setup was merged.

The right fix is to do maintenance on the GNR targets but until
that happens, just move from epoll to interrupt based testing that
is known to work in the current configuration.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>