sgx: add QuoteVerification demo and cleanup hostNetwork dependency
hostNetwork usage for SGX demo pods is not absolutely necessary so it's
better to clean it up and make IAS "security" scanners happier. It was
originally used to be able to use "localhost" PCCS but this change now
adds an example how proper PCCS url can be configured using jq.

Additionally, SGX DCAP Quote Verification is added.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>