vpu: remove deprecated plugin
The VPU plugin can only be used with devices that are
no longer supported by upper layers, such as OpenVINO.

The deprecation plan for the plugin was announced earlier
this year and post v0.28 marks the date when the plugin is removed
from the repo.

Releases before v0.29 have the plugin available should it
be needed.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>