operator: fix upgradeImages
Operatorhub bundle can have sha256 image tags that are put through
env vars. When operator controller manager gets upgraded, its
operands (plugin daemonsets) should be updated to the image in the
env vars. But it has not been working properly because of wrong
parsing.

Fix it to parse the image names that have sha256 tags correctly so
env vars in operator can be used as intended.

Signed-off-by: Hyeongju Johannes Lee <hyeongju.lee@intel.com>