qat: drop setupDeviceIDs()
setupDeviceIDs() is obsoleted and the preferred approach is driver_override
already implemented in qat-init.sh initcontainer.

The new_id mechanism was added way before we had the initcontainer support in place.
Furthermore, at least for vfio-pci we don't need it at all if the driver uses
ids=8086:<qat VF dev IDs>.

Drop write attemps to new_id in favor of the initcontainer functionality.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>