dependabot: group all github-actions updates
As we track pinned github actions, we get an update every time a new tag
appears for any action we use. Furthermore, as we have a scheduled run
on Sundays, we get many single action update PRs that have accumulated
during the week to work with on Monday mornings.

Group all github-actions updates to a single PR to help with PR merges
and to avoid unnecessary CI runs.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>