qat: add support for QAT Gen6
Linux 6.16 added the support for qat_6xxx. Update qat_plugin and NodeFeatureRules
to detect the new device (PF 4948, VF 4949).

qat_6xxx adds new service 'decomp' for decompression only operations and the
capability to configure all three services simultaneously so the plugin and
initcontainer are adapted to that.

Moreover, 'dcc' service (available on both Gen4 and Gen6) was missing so add
that too.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>