qat: handle service parsing via debugfs
The QAT out-of-tree driver does not provide sysfs based cfg_services
interface to read (and write) services configurations so QAT debugfs
is used as the fallback.

bad65eb16f changed how the services string is parsed and trimmed but
missed the debugfs path. Because of this, the resource registration errors
due so invalid resource names (e.g., "sym;dc", which should be "sym-dc").

Fix by handling the debugfs path the same way as the cfg_services sysfs
path.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>